backend/requirements.txt
mcp
mcp[cli]
langchain-ollama
langchain-community
langchain-core
pydantic
fastapi
uvicorn
fastmcp
sse-starlette
starlette
requests
httpx
python-dotenv
uvicorn[standard]
langchain_google_genai
pyautogen[ollama]
autogen-agentchat[gemini,retrievechat,lmm]
ag2[ollama,gemini]
google-genai
google-generativeai